temp_unzip/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg[binary]==3.1.18
pydantic==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.20
alembic==1.13.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.2.0
